Moved proxy bean to top level class, so it does not drag WeldFeature to native image.
