+++ integrations/graal/mp-native-image-extension/src/main/java/io/helidon/integrations/graal/mp/nativeimage/extension/WeldFeature.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2020 Oracle and/or its affiliates.
+ * Copyright (c) 2019, 2021 Oracle and/or its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,7 +17,6 @@
 package io.helidon.integrations.graal.mp.nativeimage.extension;
 
 import java.io.IOException;
-import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
 import java.net.URL;
 import java.util.ArrayList;
@@ -32,12 +31,7 @@ import java.util.Objects;
 import java.util.Set;
 import java.util.function.Supplier;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.context.spi.CreationalContext;
-import javax.enterprise.inject.Any;
-import javax.enterprise.inject.Default;
 import javax.enterprise.inject.spi.Bean;
-import javax.enterprise.inject.spi.InjectionPoint;
 import javax.json.Json;
 import javax.json.JsonArray;
 import javax.json.JsonObject;
@@ -224,76 +218,6 @@ public class WeldFeature implements Feature {
             return weldProxies;
         } catch (IOException e) {
             throw new IllegalStateException("Failed to get resources", e);
-        }
-    }
-
-    /**
-     * Proxy used to initialize Weld.
-     */
-    static final class ProxyBean implements Bean<Object> {
-        // this is the bean class (producer class, or the type itself for managed beans)
-        private final Class<?> beanClass;
-        // the types of the produced bean (or
-        private final Set<Type> types;
-
-        ProxyBean(Class<?> beanClass, Set<Type> types) {
-            this.beanClass = beanClass;
-
-            this.types = types;
-        }
-
-        @Override
-        public Class<?> getBeanClass() {
-            return beanClass;
-        }
-
-        @Override
-        public Set<InjectionPoint> getInjectionPoints() {
-            return Collections.emptySet();
-        }
-
-        @Override
-        public boolean isNullable() {
-            return false;
-        }
-
-        @Override
-        public Object create(CreationalContext<Object> creationalContext) {
-            throw new IllegalStateException("This bean should not be created");
-        }
-
-        @Override
-        public void destroy(Object instance, CreationalContext<Object> creationalContext) {
-        }
-
-        @Override
-        public Set<Type> getTypes() {
-            return types;
-        }
-
-        @Override
-        public Set<Annotation> getQualifiers() {
-            return Set.of(Any.Literal.INSTANCE, Default.Literal.INSTANCE);
-        }
-
-        @Override
-        public Class<? extends Annotation> getScope() {
-            return ApplicationScoped.class;
-        }
-
-        @Override
-        public String getName() {
-            return beanClass.getName();
-        }
-
-        @Override
-        public Set<Class<? extends Annotation>> getStereotypes() {
-            return Collections.emptySet();
-        }
-
-        @Override
-        public boolean isAlternative() {
-            return false;
         }
     }
